requirements.txt
Django
djangorestframework
django-cors-headers 
pandas
scikit-learn
vonage